requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
azure-core==1.33.0
azure-storage-blob==12.25.1
azure-storage-queue==12.15.0
bcrypt==5.0.0
certifi==2024.2.2
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.115.12
frozenlist==1.8.0
h11==0.16.0
idna==3.11
isodate==0.7.2
jwt==1.4.0
motor==3.7.1
multidict==6.7.0
postmarker==1.0
propcache==0.4.1
pyasn1==0.6.2
pycparser==2.23
pydantic==2.9.2
pydantic_core==2.23.4
pymongo==4.16.0
python-dotenv==1.2.1
python-jose==3.3.0
requests==2.32.5
rsa==4.9.1
six==1.17.0
starlette==0.46.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.34.2
yarl==1.22.0
openai==2.16.0
python-multipart==0.0.22